Transform Docker Crux (#8)
* Rename 'make.sh' to 'release' for clarity

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Use functions to break up code

As we expand the functionality of `make.sh` it should
ensure things stay neat, tidy and organised.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Make architecture detection more flexible

Use 2 methods of architecture detection in case one or the other is
not available.  Also, 'dpkg' usually refers to Debian Packaging,
which is not used by this repository.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Compact indentation and fix style issues

Using 8 spaces (or tabs) extends the lines further than they
really need to be.  Also ';;' should be on their own line
according to most examples/conventions.

This will now will save confusion as we start to chop the code
around.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Only attempt to remove cruxbuild container if it exists

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Amalgamate testing into the main build script

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* Move processes into version/architecture specific directories

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* Add LICENSE file

Taken from the upstream project:

  https://crux.nu/Main/Handbook3-4#ntoc5

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* Clean-up the process to build the CRUX root filesystem

Amalgamate the build scripts and simplify the Dockerfile.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Add usage message

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Obtain version via an argument

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Seperate committing to branch

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* crux-fetch: Standardise root filesystem filename for AArch64 and X86_64

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* Remove the media directory

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Unify commiting for all supported architectures

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Unify testing for all supported architectures

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Fix check for existing branch

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Unify building for all supported architectures

This will require a subsequent step once AArch64 has bootstrapped.

Once a CRUX image is available we'll use that as a base instead.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Silence removal of 'cruxbuild' container

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Add functionality to push committed release branch

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Seperate fetching version from command line

Not all future commands will require a version number.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* Further compress root filesystem files

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Allow ARCH to be overridden

It is possible to pre-set/export ARCH which will be detected by
the release script.  This should allow versions destined for
different architectures to be built and pushed from one machine.

Note: AArch64 is still in bootstrap mode, thus it is currently
unable to build the x86_64 images since those use 'crux:latest'
as a build platform.  This will be fixed once AArch64
'crux:latest' has been released.

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Add ability to build Docker's Official Images library file

Signed-off-by: Lee Jones <lee.jones@linaro.org>

* release: Allow any version to be pulled - not just the latest

Signed-off-by: Lee Jones <lee.jones@linaro.org>